AXIS2C-944: Fix guththila_xml_parser, guththila_xml_writer to always examine     namespace scope from the inside outwards, not just some of the time. Fix guththila_get_namespace_XXX_by_number to use consistent indexing between     validating parser and non-validating parser. Fix signed/unsigned mismatch compilation warnings. Move the guththila_token_close routines where they should be, after create/init/free.

+++ guththila/include/guththila.h
@@ -98,7 +98,7 @@ typedef struct guththila_s
 
     int guththila_event;
 
-    int next;
+    size_t next;
 
     int last_start;
